Add `Adhan-C` library for prayer times calculation
A wrapper for the library must be used as it includes a function  with the signature `is_leap_year(int year)`, this conflicts with a function in the `utz` library.
The `calendrical_helper.c` & `calendrical_helper.h` are copied from the Adhan-C library `src/` folder and modified to have a new signature: `adhan_is_leap_year(int year)`.